feature: add type file for lifecenter
