+++ src/pages/HomePage/pages/LifeCenter/LifeCenter.tsx
@@ -81,12 +81,3 @@ export function LifeCenter() {
   );
 }
 
-export type LifeCenterType = {
-  name: string;
-  description: string;
-  location: string;
-  meeting_dates: string[];
-  num_of_members: number;
-  num_of_souls_won: number;
-  id?: string;
-};
