+++ src/utils/api/lifeCenter/interface.ts
@@ -1,0 +1,9 @@
+export type LifeCenterType = {
+  name: string;
+  description: string;
+  location: string;
+  meeting_dates: string[];
+  num_of_members: number;
+  num_of_souls_won: number;
+  id?: string;
+};
